Finish using `KFunction`s instead of `Method`s
On all `Params` functions take in `KFunction`s instead of `Method`s.

With this change, the client is now able to call nullable functions
without issue. By using `KFunction`s and `KParameter`s it is possible
to determine if a type is nullable or not so when matching `null` values
the correct function is chosen.